ISPN-11072 Move initialization of Elytron security providers to a method
