+++ client/hotrod-client/src/main/java/org/infinispan/client/hotrod/RemoteCacheManager.java
@@ -8,8 +8,6 @@ import java.io.Closeable;
 import java.io.IOException;
 import java.io.InputStream;
 import java.net.SocketAddress;
-import java.security.Provider;
-import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.EnumSet;
@@ -71,7 +69,7 @@ import org.infinispan.commons.util.Version;
 import org.infinispan.counter.api.CounterManager;
 import org.infinispan.protostream.SerializationContext;
 import org.infinispan.protostream.SerializationContextInitializer;
-import org.wildfly.security.WildFlyElytronProvider;
+
 
 /**
  * <p>Factory for {@link RemoteCache}s.</p>
@@ -113,28 +111,6 @@ public class RemoteCacheManager implements RemoteCacheContainer, Closeable, Remo
    private ObjectName mbeanObjectName;
    private TimeService timeService = DefaultTimeService.INSTANCE;
    private ExecutorService asyncExecutorService;
-
-   static {
-      // Register only the providers that matter to us
-      try {
-         // Elytron >= 1.8
-         for (String name : Arrays.asList(
-               "org.wildfly.security.sasl.plain.WildFlyElytronSaslPlainProvider",
-               "org.wildfly.security.sasl.digest.WildFlyElytronSaslDigestProvider",
-               "org.wildfly.security.sasl.external.WildFlyElytronSaslExternalProvider",
-               "org.wildfly.security.sasl.oauth2.WildFlyElytronSaslOAuth2Provider",
-               "org.wildfly.security.sasl.scram.WildFlyElytronSaslScramProvider",
-               "org.wildfly.security.sasl.gssapi.WildFlyElytronSaslGssapiProvider",
-               "org.wildfly.security.sasl.gs2.WildFlyElytronSaslGs2Provider"
-         )) {
-            Provider provider = (Provider)Class.forName(name).getConstructor(new Class[]{}).newInstance(new Object[]{});
-            SecurityActions.addSecurityProvider(provider);
-         }
-      } catch (Exception e) {
-         // Elytron < 1.8
-         SecurityActions.addSecurityProvider(new WildFlyElytronProvider());
-      }
-   }
 
    /**
     * Create a new RemoteCacheManager using the supplied {@link Configuration}. The RemoteCacheManager will be started
