+++ client/hotrod-client/src/main/java/org/infinispan/client/hotrod/security/SecurityActions.java
@@ -1,14 +1,14 @@
-package org.infinispan.client.hotrod;
+package org.infinispan.client.hotrod.security;
 
 import java.security.AccessController;
 import java.security.PrivilegedAction;
 import java.security.Provider;
 
 /**
- * Privileged actions for package org.infinispan.client.hotrod
- *
- * Do not move. Do not change class and method visibility to avoid being called from other
- * {@link java.security.CodeSource}s, thus granting privilege escalation to external code.
+ * Privileged actions for package org.infinispan.client.hotrod.security
+ * <p>
+ * Do not move. Do not change class and method visibility to avoid being called from other {@link
+ * java.security.CodeSource}s, thus granting privilege escalation to external code.
  *
  * @author Scott.Stark@jboss.org
  * @since 4.2
@@ -24,6 +24,7 @@ final class SecurityActions {
 
    static void addSecurityProvider(Provider provider) {
       doPrivileged(() -> {
+
                if (java.security.Security.getProvider(provider.getName()) == null) {
                   java.security.Security.insertProviderAt(provider, 1);
                }
